+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The official release will have a defined and more stable API. If you are already
   * Added support for build-level metadata (#138). See [Build Metadata](https://github.com/flosell/lambdacd/wiki/Build%20Metadata) for details
   * UI support for some kinds of metadata: 
     * `:human-readable-build-label`
+  * Added events `:pipeline-started` and `:pipeline-finished` (#155)
 * Bug fixes:
   * Catch Exception instead of Throwable in build steps to avoid catching Errors which cannot be handled (#148), thanks @hgsy!
 * Deprecated: